Restructured `git_utils` in modules.
