+++ git_utils_shared/src/lib.rs
@@ -1,5 +1,6 @@
 use std::process::Command;
-
+pub mod embed;
+pub mod file;
 // fn run_command(command: &str, args: [&str]) -> () {}
 
 /// Get the merge base between the two provided branches
@@ -88,197 +89,4 @@ pub fn clone(git_url: &str, path: &str) {
         .status()
         .expect("Failed")
         .success();
-}
-
-pub mod embed {
-    use std::process::Command;
-
-    use itertools::Itertools;
-    pub fn add_fild_to_embed_file(project: &str, fild: &str, content: &str) {
-        let _ = Command::new("git")
-            .args([
-                "config",
-                "--file",
-                ".gitembed",
-                &format!("embed.{}.{}", project, fild),
-                content,
-            ])
-            .status()
-            .expect("Failed")
-            .success();
-    }
-
-    pub fn remove_fild_to_embed_file(project: &str, fild: &str) {
-        let _ = Command::new("git")
-            .args([
-                "config",
-                "--file",
-                ".gitembed",
-                "--unset",
-                &format!("embed.{}.{}", project, fild),
-            ])
-            .status()
-            .expect("Failed")
-            .success();
-    }
-
-    pub fn remove_section_to_embed_file(project: &str) {
-        let _ = Command::new("git")
-            .args([
-                "config",
-                "--file",
-                ".gitembed",
-                "--remove-section",
-                &format!("embed.{project}"),
-            ])
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .status()
-            .expect("Failed")
-            .success();
-    }
-
-    pub fn get_head_of_embed_project(embed_project_path: &str) -> String {
-        let stdout_raw = Command::new("git")
-            .args([
-                "--git-dir",
-                &format!("{}/.egit", embed_project_path),
-                "rev-parse",
-                "HEAD",
-            ])
-            .output()
-            .expect("Failed")
-            .stdout;
-        String::from_utf8(stdout_raw).unwrap().trim().to_string()
-    }
-
-    #[derive(Debug)]
-    pub struct EmbedEnty {
-        pub name: String,
-        pub git_url: String,
-        pub path: String,
-        pub head: String,
-    }
-    pub fn get_embeds() -> Vec<EmbedEnty> {
-        let stdout_raw = Command::new("git")
-            .args(["config", "--file", ".gitembed", "--list"])
-            .output()
-            .expect("Failed")
-            .stdout;
-        let stdout = String::from_utf8(stdout_raw).unwrap().trim().to_string();
-
-        let mut embeds = Vec::new();
-
-        for (url_line, path_line, head_line) in stdout.lines().tuples() {
-            let get_value = |line: &str| -> String {
-                line.split("=")
-                    .collect::<Vec<&str>>()
-                    .last()
-                    .unwrap()
-                    .to_string()
-            };
-
-            let git_url = get_value(url_line);
-            let path = get_value(path_line);
-            let head = get_value(head_line);
-
-            let name = path_line
-                .split("=")
-                .collect::<Vec<&str>>()
-                .first()
-                .unwrap()
-                .split(".")
-                .collect::<Vec<&str>>()
-                .get(1)
-                .unwrap()
-                .to_string();
-
-            embeds.push(EmbedEnty {
-                name,
-                git_url,
-                path,
-                head,
-            });
-        }
-        embeds
-    }
-}
-pub mod file {
-    #[derive(Hash, Eq, PartialEq, Clone, Debug)]
-    pub struct GitStatusFile {
-        name: String,
-        status: GitStatus,
-    }
-    #[derive(Hash, Eq, PartialEq, Clone, Debug)]
-    pub enum GitStatus {
-        Renamed,
-        Modified,
-        Deleted,
-        Added,
-        Untracked,
-    }
-    use std::{collections::HashMap, process::Command};
-
-    pub fn git_status() -> HashMap<GitStatus, GitStatusFile> {
-        let cmd = Command::new("git")
-            .args(["status", "--porcelain=v2", "-s"])
-            .output()
-            .expect("Failed ")
-            .stdout;
-        let stdout = String::from_utf8(cmd).unwrap();
-        let mut git_status_record = HashMap::new();
-        for line in stdout.split("\n") {
-            let line_split: Vec<&str> = line.trim().split(" ").filter(|c| !c.is_empty()).collect();
-
-            dbg!(&line_split);
-
-            use GitStatus::*;
-            let status = match line_split[0] {
-                "AD" | "D" => Deleted,
-                "AM" | "M" => Modified,
-                "AR" | "R" => Renamed,
-                "A" => Added,
-                "??" => Untracked,
-                err => panic!("{} is not an implemented git status", err),
-            };
-
-            git_status_record.insert(
-                status.clone(),
-                GitStatusFile {
-                    name: line_split[1].to_string(),
-                    status: status,
-                },
-            );
-        }
-        git_status_record
-    }
-    pub fn modified_files(file_paths: &Vec<String>) -> Option<Vec<String>> {
-        let stdout_raw = Command::new("git")
-            .args(["--no-pager", "diff", "--name-only", "-z"])
-            .args(file_paths)
-            .output()
-            .expect("Failed ")
-            .stdout;
-        let stdout_str = String::from_utf8(stdout_raw).unwrap();
-
-        if stdout_str.is_empty() {
-            return None;
-        }
-        let files = stdout_str[..stdout_str.len() - 1]
-            .split("\0")
-            .map(|path| path.to_owned())
-            .collect::<Vec<String>>();
-        Some(files)
-    }
-
-    pub fn files_are_tracked(file_paths: &Vec<String>) -> bool {
-        Command::new("git")
-            .args(["ls-files", "--error-unmatch"])
-            .args(file_paths)
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .status()
-            .expect("Failed ")
-            .success()
-    }
 }
